Moved redeploy logic into own class
