+++ src/test/groovy/no/skatteetaten/aurora/boober/service/AbstractMockedOpenShiftSpecification.groovy
@@ -49,7 +49,8 @@ import spock.mock.DetachedMockFactory
     AuroraDeploymentSpecValidator,
     SharedSecretReader,
     VelocityTemplateJsonService,
-    OpenShiftObjectLabelService
+    OpenShiftObjectLabelService,
+    RedeployService
 ])
 class AbstractMockedOpenShiftSpecification extends AbstractSpec {
